added configMap file and cleaned up requestReducerOnline and requestReducerOffline Co-authored-by: Eric Gomez <ergomez0201@users.noreply.github.com>
