+++ disco-db-demo/public/swCacheSite-indexedDB.js
@@ -1,18 +1,57 @@
 //public/sw.js
 //importScripts('https://cdn.jsdelivr.net/npm/dexie@3.2.1/dist/dexie.min.js');
-import { openDB, dbAdd, dbDeleteAll, addToSyncQueue, syncDataToServer, dbGetAll, dbDeleteOne, dbUpdateOne, dbGlobals } from './indexedDB.js';
+import { openDB, syncDataToServer, dbGlobals } from './indexedDB.js';
+import { requestReducerOffline } from './requestReducerOffline.js';
+import { requestReducerOnline } from './requestReducerOnline.js';
+
 
 
 const cacheName = 'my-site-cache-v3';
+// do we need to store DB in a global variable?
 
-let DB;
+const config = {
+  onlineRoutes: [ 
+    { url: 'http://localhost:3000/user/load',
+      store : 'notesStore' 
+    },
+  ],
+  offlineRoutes: [ 
+    { url: 'http://localhost:3000/user/load',
+      store : 'notesStore' 
+    },
+    { url: 'http://localhost:3000/user/notes',
+      store: 'notesStore'
+    },
+  ]
+}
+
+// using map
+const onlineUrlStoreMap = new Map();
+config.onlineRoutes.forEach(el => {
+  onlineUrlStoreMap.set(el.url, el.store);
+});
+
+const offlineUrlStoreMap = new Map();
+config.offlineRoutes.forEach(el => {
+  offlineUrlStoreMap.set(el.url, el.store);
+});
+
+// const urlArr = [];
+// config.routes.forEach(el => {
+//   urlArr.push(el.url);
+// });
+
+// const storeArr = [];
+// config.routes.forEach(el => {
+//   storeArr.push(el.store);
+// });
 
 
 self.addEventListener('install', event => {
   console.log('Attempting to install service worker and cache static assets');
   self.skipWaiting();
   console.log('opening DB since sw is activated')
-        openDB().then(data => DB = data);
+       openDB();
 });
 
 
@@ -58,19 +97,24 @@ self.addEventListener('fetch', event => {
         .then(cache => {
           //Add response to cache
           cache.put(event.request, resCloneCache);
-        })
-      // if(url is in url.config) 
-      //    func(event.request.method, event.request.url, data)
-      
+        })      
       //invoke online reducer to populate indexedDB
-      requestReducerOnline(method, url, event.request, resCloneDB);
+      if (onlineUrlStoreMap.has(url)){
+        console.log('onlineUrlStoreMap.get(url): ', onlineUrlStoreMap.get(url))
+        console.log('intercepting URL: ' + url, ' and method: ', method)
+        requestReducerOnline(method, url, onlineUrlStoreMap.get(url), resCloneDB); // Eric: figure out how best to pass store into this reducer. Match parameters to the offline reducer. Look at catch block 
+      }
       return response;
     })
     // if network is unavailable
     .catch((err) => {
-      console.log('this is DB in catch block: ', DB);
+      console.log('this is DB in catch block: ', dbGlobals.DB);
       //invoke offline reducer to perform RUD functions to indexedDB
-      return requestReducerOffline(method, url, reqClone);
+      // if (urlArr.includes(url))
+      if (offlineUrlStoreMap.has(url)){
+        console.log('offlineUrlStoreMap.get(url): ', offlineUrlStoreMap.get(url))
+        return requestReducerOffline(method, url, offlineUrlStoreMap.get(url), reqClone); // Eric: figure out how best to pass store into this reducer
+      }
     })
   )
 });
@@ -99,163 +143,130 @@ function backgroundSync() {
 }
 
 
-//create a request reducer - online
-function requestReducerOnline(method, url, eventRequest, clonedResponse) {
-  switch(url) {
-    case 'http://localhost:3000/user/load':
-      switch(method) {
-        case 'GET':
-          const resCloneDB = clonedResponse;
-          resCloneDB.json().then(data => {
-            console.log('this is the rescloneDB: ', data)
-            //delete existing indexedDB data
-            if (DB) {
-              dbDeleteAll();
-            } else {
-              openDB( () => {
-                dbDeleteAll();
-              })
-            }
-            //populate indexedDB here
-            data.data.forEach( note => {
-              console.log('this is the note object: ', note);
-              if (DB) {
-                dbAdd(note);
-              } else {
-                openDB( () => {
-                  dbAdd(note);
-                })
-              }
-            })
-            console.log('returning eventresponse after adding all notes into IndexedDB');
-          });
+// //create a request reducer - online
+// function requestReducerOnline(method, url, eventRequest, clonedResponse) {
+//   switch(url) {
+//     case 'http://localhost:3000/user/load':
+//       switch(method) {
+//         case 'GET':
+//           const resCloneDB = clonedResponse;
+//           resCloneDB.json().then(data => {
+//             console.log('this is the rescloneDB: ', data)
+//             //delete existing indexedDB data
+//             if (DB) {
+//               dbDeleteAll();
+//             } else {
+//               openDB( () => {
+//                 dbDeleteAll();
+//               })
+//             }
+//             //populate indexedDB here
+//             data.data.forEach( note => {
+//               console.log('this is the note object: ', note);
+//               if (DB) {
+//                 dbAdd(note);
+//               } else {
+//                 openDB( () => {
+//                   dbAdd(note);
+//                 })
+//               }
+//             })
+//             console.log('returning eventresponse after adding all notes into IndexedDB');
+//           });
 
-        default:
-          console.log('this method is not configured');
-          break;
-      }
-    default:
-      console.log('this is the online eventResponse: ', eventRequest);
-      break;
-  }
-}
-
-// if (url === x && method: GET)
-
-//url
-  //GET
-    //requestReducerOfflineGET(url)
-  //DELETE
-    //requestReducerOfflineDELETE(url)
-  //PATCH
-    //requestReducerOfflinePATCH(url)
-
-//config object
-// {
-// url:'',
-// method:''
+//         default:
+//           console.log('this method is not configured');
+//           break;
+//       }
+//     default:
+//       console.log('this is the online eventResponse: ', eventRequest);
+//       break;
+//   }
 // }
 
-// {routes: 
-//   { url: '',
-//     store : [] 
-//   },
-//   { url: '',
-//   methods: []
-//   },
+// //create a request reducer - offline
+// function requestReducerOffline(method, url, eventRequest, eventResponse) {
+//   switch(url) {
+//     case 'http://localhost:3000/user/load':
+//       switch(method) {
+//         case 'GET':
+//           if (DB) {
+//             return dbGetAll().then((data) => {
+//               //REVISIT THIS, make sure to change data back to data!!
+//               const responseBody = { data };
+//               console.log('this is the response body inside the request reducer function: ');
+//               console.log({responseBody});
+//               const IDBData = new Response(JSON.stringify(responseBody));
+//               return IDBData;
+//             })
+//           } else {
+//             return openDB( () => {
+//               console.log('invoking dbGetAll in else')
+//               dbGetAll().then((data) => {
+//                 const responseBody = {data: data};
+//                 const IDBData = new Response(JSON.stringify(responseBody));
+//                 return IDBData;
+//               });
+//             })
+//           }
+//         default:
+//           console.log('this method is not configured');
+//           break;
+//       }
+//     case 'http://localhost:3000/user/notes':
+//       switch(method) {
+//         case 'DELETE':
+//           return eventRequest.json()
+//           .then((data) => {
+//             const reqBody = {
+//               url: url,
+//               method: method,
+//               body: data
+//             };
+//             console.log('this is the delete data object when network fails: ', data);
+//             backgroundSync();
+//             addToSyncQueue(reqBody);
+
+//             const id = data._id;
+//             //call function to DELETE note
+//             dbDeleteOne(id);
+//             const deleteResponse = new Response(JSON.stringify({}));
+//             console.log({ deleteResponse });
+//             return deleteResponse;
+//           })
+//           .catch( err => {
+//             console.log('this is in the dbDeleteOne catch block: ', err);
+//           })
+//         case 'PATCH':
+//           return eventRequest.json()
+//           .then((data) => {
+//             const reqBody = {
+//               url: url,
+//               method: method,
+//               body: data
+//             };
+//             backgroundSync();
+//             addToSyncQueue(reqBody);
+//             //call function to UPDATE note
+//             const id = data._id;
+//             console.log('this is the data sent to dbUpdateOne: ', data);
+//             dbUpdateOne(data);
+
+//             // returns empty object to trigger rerender in our app 
+//             // assumes developer does not want to do anything with the response
+//             const patchResponse = new Response(JSON.stringify({}));
+//             console.log({ patchResponse });
+//             return patchResponse;
+//           })
+//         default:
+//           console.log('this method in /user/notes is not configured')
+//       } 
+//     default:
+//       console.log('this url is not configured');
+//       return caches.match(eventRequest)
+//         .then(response => {
+//           console.log('-----------this is in the caches response block: ', eventRequest);
+//           return response
+//         })
+//   }
 // }
-
-//      GET method at URL
-// else if (url === x && method: DELETE)
-// function requestReducerOfflineDELETE(url, data)
-//      DELETE method at URL
-// else if (url === x && method: PATCH)
-// function requestReducerOfflinePATCH(url, data)
-//      PATCH method at URL
-
-//create a request reducer - offline
-function requestReducerOffline(method, url, eventRequest, eventResponse) {
-  switch(url) {
-    case 'http://localhost:3000/user/load':
-      switch(method) {
-        case 'GET':
-          if (DB) {
-            return dbGetAll().then((data) => {
-              //REVISIT THIS, make sure to change data back to data!!
-              const responseBody = { data };
-              console.log('this is the response body inside the request reducer function: ');
-              console.log({responseBody});
-              const IDBData = new Response(JSON.stringify(responseBody));
-              return IDBData;
-            })
-          } else {
-            return openDB( () => {
-              console.log('invoking dbGetAll in else')
-              dbGetAll().then((data) => {
-                const responseBody = {data: data};
-                const IDBData = new Response(JSON.stringify(responseBody));
-                return IDBData;
-              });
-            })
-          }
-        default:
-          console.log('this method is not configured');
-          break;
-      }
-    case 'http://localhost:3000/user/notes':
-      switch(method) {
-        case 'DELETE':
-          return eventRequest.json()
-          .then((data) => {
-            const reqBody = {
-              url: url,
-              method: method,
-              body: data
-            };
-            console.log('this is the delete data object when network fails: ', data);
-            backgroundSync();
-            addToSyncQueue(reqBody);
-
-            const id = data._id;
-            //call function to DELETE note
-            dbDeleteOne(id);
-            const deleteResponse = new Response(JSON.stringify({}));
-            console.log({ deleteResponse });
-            return deleteResponse;
-          })
-          .catch( err => {
-            console.log('this is in the dbDeleteOne catch block: ', err);
-          })
-        case 'PATCH':
-          return eventRequest.json()
-          .then((data) => {
-            const reqBody = {
-              url: url,
-              method: method,
-              body: data
-            };
-            backgroundSync();
-            addToSyncQueue(reqBody);
-            //call function to UPDATE note
-            const id = data._id;
-            console.log('this is the data sent to dbUpdateOne: ', data);
-            dbUpdateOne(data);
-
-            // returns empty object to trigger rerender in our app 
-            // assumes developer does not want to do anything with the response
-            const patchResponse = new Response(JSON.stringify({}));
-            console.log({ patchResponse });
-            return patchResponse;
-          })
-        default:
-          console.log('this method in /user/notes is not configured')
-      } 
-    default:
-      console.log('this url is not configured');
-      return caches.match(eventRequest)
-        .then(response => {
-          console.log('-----------this is in the caches response block: ', eventRequest);
-          return response
-        })
-  }
-}
